Countdown and redirect working
When the tab changes the background script checks if the new tab is on
the restricted list and then figures out how much time is left if it
is. Also, when the tab is changed the remaining time in the last
category is updated if the last page was on the restricted site.